Finished GameBoard class
GameBoard class is finished. It mimics the gameboard that you'd make on paper, along with the information that's contained there: hits, misses, all_spaces(that are available). and a means of displaying the board. Future things: making a dict? that with position tuple as key and display info as value.